Add files via upload
add_course: adds a completed course to the information of a specific student in the database. The course data is a tuple consisting of the name of the course and the grade. Courses with grade 0 is ignored when adding course information. Additionally, if the course is already in the database in that specific student's information, the grade recorded in the database is not lowered if the course is repeated.